requirements.txt
Django>=4.2,<5.0
django-import-export>=3.3.0
django-simple-captcha>=0.6.0
APScheduler>=3.10.0
requests>=2.31.0
openpyxl>=3.1.0
